:recycle: refactor(User): - User Service Command/Query 분리 - Quiz, Coin Service에 User 누적 정보 로직 추가 - User Update 로직 수정 (#4)

+++ src/main/java/com/sparta/devquiz/domain/quiz/entity/UserQuiz.java
@@ -34,5 +34,5 @@ public class UserQuiz extends BaseTimeEntity {
     private UserQuizStatus status;
 
     @Column(nullable = false)
-    private Long score;
+    private int score;
 }
